REFACTOR: NAV-47 - Extract route scanner from raptor
